Updated documentation.
Added sphinx-rtd theme and watchdog for better doc writing.
Changed all .md files to .rst files because RTD supports it
a little bit better, and mostly just to be consistent. Removed
the old docs in place of a more full-featured set of docs.

Cleaned up pydoc documentation as some of it didn't look quite
right in the new theme. So that is all the changes.