Issue/27 (#30)
* Added convinence method to errors module.

This will replace the functionality in plugin later. For now
it is handy to be able to convert any arbitrary python exception
into a JSON string.

* Added is_json method to request.

* Better JSON Exception parsing. Solves #27.

This will handle JSON exceptions better. You can now raise
exceptions in plugins the following ways and it will look
good in the GUI.

```python
def my_command(self):
    raise Exception({"foo": "bar"})
    raise Exception(json.dumps({"foo": "bar"})
    raise Exception("message")
```

Each will parse into valid JSON.

* Fixed flake8

* Fixed tests for python2.

Added a note to the new method. Apparently, in python 2, custom
attributes on exception are not part of the args of an exception.
There really isn't a good way around this, so the tests are for
the particular version of python.

* Updated documentation.