Added trigger types as models.
This allows us to more accurately define all of the triggers
available in Beer Garden. This required including a new dependency
called marshmallow-polyfield:

* https://github.com/Bachmann1234/marshmallow-polyfield

This allows us to utilize our schema the same way, but we can
dynamically generate the correct type when parsing. If we didnt
want to add an extra dependency, we could just copy this library
as it is basically just one line.

There are some refactored unit tests here. I'm slowly making my
way through the code base and converting over to pytest.